requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-rest-knox==5.0.2
django-simple-history==3.7.0
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.55.2
identify==2.6.3
idna==3.10
isort==5.13.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
nodeenv==1.9.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10  # Make sure, the folder which contains `pg_config` exe, is added to PATH.
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
